refactored application.py into lots of files within the app directory
